35.0
so i was working om the update to the search engine, but I realized that there is a problem with the color and icons. It is rather illogical how I have to write out the color and src to image, when he program could do it for me. So i'm trying to implement this into the packs. If it works, I'll continue to do so in "zeleny_kalendar"