49.13
added a div regarding the politics and how we are taking care of the enviromental crisis. We have two unions + a state (slovakia, european union and united nations). Later on I'll add info about what is the newest enviro thing they are working on plus the institutions that are assigned to these cases.